refactoring files using sentient #35
adds further deatils of the aim of the repo;
* renames main source path and test scripts and notebooks;
