+++ dependencies/README.md
@@ -7,7 +7,7 @@ Install [mamba](https://github.com/mxochicale/code/tree/main/mamba) and create [
 
 * Python package versions (log for machine with CUDA GPU)
 ```
-mamba activate rtt4ssaVE
+mamba activate sentientVE
 python packages_versions.py
