+++ dependencies/no-gpu-ve.yml
@@ -12,7 +12,7 @@
 ## import torch
 ## torch.cuda.is_available()
 
-name: rtt4ssaVE
+name: sentientVE
 channels:
   - conda-forge # notebook; jupyter_contrib_nbextensions; tqdm; monai; nibabel; gdown; imgaug; ipywidgets; PyAV
 dependencies:
