+++ dependencies/ve.yml
@@ -12,7 +12,7 @@
 ## import torch
 ## torch.cuda.is_available()
 
-name: rtt4ssaVE
+name: sentientVE
 channels:
   - conda-forge 
 dependencies:
@@ -26,7 +26,6 @@ dependencies:
      - matplotlib
      - numpy
      - jupyter
-     #- jupyter_contrib_nbextensions #>>> jupyter nbextensions_configurator enable
      - notebook
      - pandas
      - seaborn
