--- rtt4ssa/README.md
@@ -1,8 +1,0 @@
-# rtt4ssa
-
-## Running notebooks
-```
-cd $HOME/repositories/rtt4ssa/rtt4ssa
-export PYTHONPATH="${PYTHONPATH}:$HOME/repositories/rtt4ssa"
-mamba activate rtt4ssaVE && jupyter notebook --browser=firefox
-```
