--- rtt4ssa/utils/__init__.py
@@ -1,1 +1,0 @@
-from rtt4ssa.utils.utils import *
