+++ sentient/README.md
@@ -1,0 +1,8 @@
+# sentient
+
+## Running notebooks
+```
+cd $HOME/repositories/sentient/sentient
+export PYTHONPATH="${PYTHONPATH}:$HOME/repositories/sentient"
+mamba activate sentientVE && jupyter notebook --browser=firefox
+```
