+++ sentient/video_devices/capturing_video.py
