+++ sentient/video_devices/simple_record_video.py
